Fix rust book function to work on 2-digit chapters
Instead of taking the chapter and section as one string, take each
separately and format them to add the leading zero if necessary.  This
is necessary because the tags do not have a leading zero and the files
do.
Furthermore, run $current_chapter through head to avoid opening the
entire directory in firefox.