Use <<- to indent here-document then fix tabs
Have to use tabs with <<- and the terminal renders tabs as 8 columns
wide by default so set that to the correct value with tabs.  Then less
doesn't listen to what tabs does, so have to fix that separately.  Then
git normally steps on $LESS but won't if it's set, so it does nothing
and diffs come out like garbage, so have to fix that.  Maybe we should
include FRX in $LESS too, but I don't know.  Pain.